fix: update iconUrl and add EASTER_AVATAR constant
This commit updates the `iconUrl` property in the `Create` class of the `hub/create.ts` file to use the `EASTER_AVATAR` constant from `Constants.ts` if `iconUrl` is not provided. Additionally, a new constant `EASTER_AVATAR` is added to `Constants.ts` with the value `'https://i.imgur.com/80nqtSg.png'`.