refactor: handleComponents method to static in command classes and some deepsource hotfixes (#69)
* refactor: some deepsource hotfixes ig

* refactor: Fix BlockNSFW setting in Manage command

This commit fixes an issue in the Manage command where the BlockNSFW setting was not working correctly. The code now properly handles the setting and displays an error message if it cannot be changed yet. This resolves a bug reported by users.

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>

* refactor: use template literals instead of string concat

* refactor: execute string and channel selects in Connection class

This commit refactors the Connection class in the slash/Main directory. It introduces two new methods, executeStringSelects and executeChannelSelects, to handle string and channel select menu interactions respectively. This improves code organization and readability. The changes also include updating the method names in the code to match the new method names.

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>

* idk

---------

Co-authored-by: deepsource-autofix[bot] <62050782+deepsource-autofix[bot]@users.noreply.github.com>