feat: Add progress spinner to command registration (#156)
The changes in this commit add a progress spinner to the command registration process in the `deploy-commands.js` script. This is done to provide a more user-friendly experience while the commands are being registered, as the registration process can take some time, especially for a large number of commands.

The key changes include:

- Introducing a `startSpinner` function that displays a spinning progress indicator with the total number of commands being registered.
- Clearing the spinner line and printing a "Done!" message when the registration is complete.
- Handling the case where the build is not yet completed by throwing an error instead of exiting the process.

These changes improve the overall user experience and provide better visual feedback during the command registration process.