refactor: remove redudant jumps
Updates the logic for handling the connection embed color to use
the nullish coalescing operator to set the value to `{ unset: true }`
if no color is provided.

fix(hub): remove unnecessary return statement

Removes the unnecessary `return` statement at the end of the
`BlockWordCommand` class, as the function already returns a reply
to the user.